requirements310.txt
pyinstaller==6.9.0
flask
Werkzeug==2.2.2
gevent
psutil
requests
pybase64
websockets
websocket-client
pillow==9.3.0
lxml
rsa
requests_toolbelt
pystray
pyyaml
openpyxl
aiohttp
qrcode
brotli
pyjson5
APScheduler==3.10.1
js2py
certifi
httpx
prompt_toolkit
regex
rich
filetype
grpcio
grpcio-tools
protobuf
